optimization of speakers module - removed unnecessary speaker file
